Some changes to CDInfo, make both CDInfoPrivate and TrackInfoPrivate inherit from the same class, so both can use the same methods. Used for creating lines, escaping/unescaping text and getting/setting data.
So now the enums can be used in get/set for both CDInfo and TrackInfo.

Also don't expose trackInfoList, but instead add methods to access tracks
and create them if they don't exist yet.

Adapt other files to the new api, and add a new test for reading/writing
CDInfo to/from text

svn path=/trunk/KDE/kdemultimedia/libkcddb/; revision=492867