Use exported targets the right way.
Besides the exported targes file and the Config file having different
purposes, so far applications wishing to use libkcddb (such as kscd)
had to use the full imported target name (which was weirdly named
`KDE::Libkcddbkcddb') instead of a variable such as
`LIBKCDDB_LIBRARIES'.

We now have a LibkcddbConfig.cmake.in file which is used to generate a
proper Config file. It currently defines the `LIBKCDDB_INCLUDE_DIRS'
and `LIBKCDDB_LIBRARIES' variables and includes the exported targets
file, which is now named LibkcddbTargets.cmake.

Reviewed-by:   Alexander Neundorf
REVIEW:        104544