Use relative includes for own public headers & export macro header
The headers are in the same folder both in the sources & the install dir,
and the export header is in the same install dir as well.
So relying on the POSIX standard to first check the same dir with ""
includes speeds up the header find (and also ensures own headers are
found first).

GIT_SILENT